requirements.txt
flask==3.0.0
opencv-python==4.8.1
numpy==1.24.3
joblib==1.3.2
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
tqdm==4.66.1 